LayoutTests/fast/css/pseudo-default-001.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
<style>
 input:default { background: lime; }
 button:default { background: lime; }
 input { background: red; }
 button { background: red; }
</style>
</head>
<body>
<p id="description"></p>
<form method="get">
<input name="victim" type="submit" value="Submit" />
</form>
<div id="console"></div>
<script>
description("This test performs a simple check for the :default CSS selector.");

v = document.getElementsByName("victim");
for (i = 0; i < v.length; i++)
    shouldBe("document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color')", "'rgb(0, 255, 0)'");

var successfullyParsed = true;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
